Request remote ArcGIS resources so they are processed
File formats on ArcGIS are not automatically available. They are
generated on the fly the first time they are requested. To avoid
problems with the previews, when the harvester is parsing the data.json
file, any suitable URL will be pinged with a HEAD request. The relevant
resource will store `requested=True` so the url is not pinged again on
successive harvestings.